requirements.txt
fastapi
uvicorn
websockets
loguru
openai
anthropic
streamlit
requests
python-dotenv
beautifulsoup4